requirements.txt
amqp==5.3.1
asgiref==3.11.1
billiard==4.2.4
celery==5.3.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==4.2.9
django-cors-headers==4.3.1
django-debug-toolbar==6.2.0
django-redis==5.4.0
djangorestframework==3.14.0
Faker==22.0.0
iniconfig==2.3.0
kombu==5.6.2
packaging==26.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg==3.2.3
psycopg-binary==3.2.3
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.0.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
vine==5.1.0
wcwidth==0.6.0
